+++ code/parachain/runtime/composable/src/weights/mod.rs
@@ -4,7 +4,6 @@ pub mod assets_registry;
 pub mod balances;
 pub mod collator_selection;
 pub mod collective;
-pub mod conviction_voting;
 pub mod frame_system;
 pub mod indices;
 pub mod membership;
@@ -12,11 +11,9 @@ pub mod multisig;
 pub mod pallet_ibc;
 pub mod pallet_proxy;
 pub mod pallet_xcm;
-pub mod referenda;
 pub mod scheduler;
 pub mod session;
 pub mod timestamp;
 pub mod tokens;
 pub mod treasury;
 pub mod utility;
-pub mod whitelist;
