+++ code/parachain/runtime/picasso/Cargo.toml
@@ -175,9 +175,6 @@ runtime-benchmarks = [
   "oracle/runtime-benchmarks",
   "pablo/runtime-benchmarks",
   "pallet-assets/runtime-benchmarks",
-  "pallet-referenda/runtime-benchmarks",
-  "pallet-conviction-voting/runtime-benchmarks",
-  "pallet-whitelist/runtime-benchmarks",
   "pallet-xcm/runtime-benchmarks",
   "proxy/runtime-benchmarks",
   "scheduler/runtime-benchmarks",
