+++ code/parachain/runtime/picasso/src/weights/mod.rs
@@ -4,7 +4,6 @@ pub mod assets_registry;
 pub mod balances;
 pub mod collator_selection;
 pub mod collective;
-pub mod conviction_voting;
 pub mod crowdloan_rewards;
 pub mod frame_system;
 pub mod identity;
@@ -16,7 +15,6 @@ pub mod pablo;
 pub mod pallet_ibc;
 pub mod pallet_xcm;
 pub mod proxy;
-pub mod referenda;
 pub mod scheduler;
 pub mod session;
 pub mod timestamp;
@@ -24,4 +22,3 @@ pub mod tokens;
 pub mod treasury;
 pub mod utility;
 pub mod vesting;
-pub mod whitelist;
