--- code/parachain/runtime/picasso/src/weights/referenda.rs
@@ -1,483 +1,0 @@
-
-#![cfg_attr(rustfmt, rustfmt_skip)]
-#![allow(unused_parens)]
-#![allow(unused_imports)]
-#![allow(missing_docs)]
-
-use frame_support::{traits::Get, weights::Weight};
-use core::marker::PhantomData;
-
-/// Weight functions for `pallet_referenda`.
-pub struct WeightInfo<T>(PhantomData<T>);
-impl<T: frame_system::Config> pallet_referenda::WeightInfo for WeightInfo<T> {
-	/// Storage: Referenda ReferendumCount (r:1 w:1)
-	/// Proof: Referenda ReferendumCount (max_values: Some(1), max_size: Some(4), added: 499, mode: MaxEncodedLen)
-	/// Storage: Scheduler Agenda (r:1 w:1)
-	/// Proof: Scheduler Agenda (max_values: None, max_size: Some(38963), added: 41438, mode: MaxEncodedLen)
-	/// Storage: Referenda ReferendumInfoFor (r:0 w:1)
-	/// Proof: Referenda ReferendumInfoFor (max_values: None, max_size: Some(936), added: 3411, mode: MaxEncodedLen)
-	fn submit() -> Weight {
-		// Proof Size summary in bytes:
-		//  Measured:  `291`
-		//  Estimated: `42428`
-		// Minimum execution time: 40_432_000 picoseconds.
-		Weight::from_parts(41_423_000, 0)
-			.saturating_add(Weight::from_parts(0, 42428))
-			.saturating_add(T::DbWeight::get().reads(2))
-			.saturating_add(T::DbWeight::get().writes(3))
-	}
-	/// Storage: Referenda ReferendumInfoFor (r:1 w:1)
-	/// Proof: Referenda ReferendumInfoFor (max_values: None, max_size: Some(936), added: 3411, mode: MaxEncodedLen)
-	/// Storage: Scheduler Agenda (r:2 w:2)
-	/// Proof: Scheduler Agenda (max_values: None, max_size: Some(38963), added: 41438, mode: MaxEncodedLen)
-	fn place_decision_deposit_preparing() -> Weight {
-		// Proof Size summary in bytes:
-		//  Measured:  `544`
-		//  Estimated: `83866`
-		// Minimum execution time: 52_009_000 picoseconds.
-		Weight::from_parts(54_126_000, 0)
-			.saturating_add(Weight::from_parts(0, 83866))
-			.saturating_add(T::DbWeight::get().reads(3))
-			.saturating_add(T::DbWeight::get().writes(3))
-	}
-	/// Storage: Referenda ReferendumInfoFor (r:1 w:1)
-	/// Proof: Referenda ReferendumInfoFor (max_values: None, max_size: Some(936), added: 3411, mode: MaxEncodedLen)
-	/// Storage: Referenda DecidingCount (r:1 w:0)
-	/// Proof: Referenda DecidingCount (max_values: None, max_size: Some(14), added: 2489, mode: MaxEncodedLen)
-	/// Storage: Referenda TrackQueue (r:1 w:1)
-	/// Proof: Referenda TrackQueue (max_values: None, max_size: Some(2012), added: 4487, mode: MaxEncodedLen)
-	/// Storage: Scheduler Agenda (r:1 w:1)
-	/// Proof: Scheduler Agenda (max_values: None, max_size: Some(38963), added: 41438, mode: MaxEncodedLen)
-	fn place_decision_deposit_queued() -> Weight {
-		// Proof Size summary in bytes:
-		//  Measured:  `3331`
-		//  Estimated: `42428`
-		// Minimum execution time: 69_077_000 picoseconds.
-		Weight::from_parts(71_533_000, 0)
-			.saturating_add(Weight::from_parts(0, 42428))
-			.saturating_add(T::DbWeight::get().reads(4))
-			.saturating_add(T::DbWeight::get().writes(3))
-	}
-	/// Storage: Referenda ReferendumInfoFor (r:1 w:1)
-	/// Proof: Referenda ReferendumInfoFor (max_values: None, max_size: Some(936), added: 3411, mode: MaxEncodedLen)
-	/// Storage: Referenda DecidingCount (r:1 w:0)
-	/// Proof: Referenda DecidingCount (max_values: None, max_size: Some(14), added: 2489, mode: MaxEncodedLen)
-	/// Storage: Referenda TrackQueue (r:1 w:1)
-	/// Proof: Referenda TrackQueue (max_values: None, max_size: Some(2012), added: 4487, mode: MaxEncodedLen)
-	/// Storage: Scheduler Agenda (r:1 w:1)
-	/// Proof: Scheduler Agenda (max_values: None, max_size: Some(38963), added: 41438, mode: MaxEncodedLen)
-	fn place_decision_deposit_not_queued() -> Weight {
-		// Proof Size summary in bytes:
-		//  Measured:  `3351`
-		//  Estimated: `42428`
-		// Minimum execution time: 68_115_000 picoseconds.
-		Weight::from_parts(70_485_000, 0)
-			.saturating_add(Weight::from_parts(0, 42428))
-			.saturating_add(T::DbWeight::get().reads(4))
-			.saturating_add(T::DbWeight::get().writes(3))
-	}
-	/// Storage: Referenda ReferendumInfoFor (r:1 w:1)
-	/// Proof: Referenda ReferendumInfoFor (max_values: None, max_size: Some(936), added: 3411, mode: MaxEncodedLen)
-	/// Storage: Referenda DecidingCount (r:1 w:1)
-	/// Proof: Referenda DecidingCount (max_values: None, max_size: Some(14), added: 2489, mode: MaxEncodedLen)
-	/// Storage: Balances InactiveIssuance (r:1 w:0)
-	/// Proof: Balances InactiveIssuance (max_values: Some(1), max_size: Some(16), added: 511, mode: MaxEncodedLen)
-	/// Storage: Scheduler Agenda (r:2 w:2)
-	/// Proof: Scheduler Agenda (max_values: None, max_size: Some(38963), added: 41438, mode: MaxEncodedLen)
-	fn place_decision_deposit_passing() -> Weight {
-		// Proof Size summary in bytes:
-		//  Measured:  `544`
-		//  Estimated: `83866`
-		// Minimum execution time: 64_860_000 picoseconds.
-		Weight::from_parts(66_772_000, 0)
-			.saturating_add(Weight::from_parts(0, 83866))
-			.saturating_add(T::DbWeight::get().reads(5))
-			.saturating_add(T::DbWeight::get().writes(4))
-	}
-	/// Storage: Referenda ReferendumInfoFor (r:1 w:1)
-	/// Proof: Referenda ReferendumInfoFor (max_values: None, max_size: Some(936), added: 3411, mode: MaxEncodedLen)
-	/// Storage: Referenda DecidingCount (r:1 w:1)
-	/// Proof: Referenda DecidingCount (max_values: None, max_size: Some(14), added: 2489, mode: MaxEncodedLen)
-	/// Storage: Balances InactiveIssuance (r:1 w:0)
-	/// Proof: Balances InactiveIssuance (max_values: Some(1), max_size: Some(16), added: 511, mode: MaxEncodedLen)
-	/// Storage: Scheduler Agenda (r:2 w:2)
-	/// Proof: Scheduler Agenda (max_values: None, max_size: Some(38963), added: 41438, mode: MaxEncodedLen)
-	fn place_decision_deposit_failing() -> Weight {
-		// Proof Size summary in bytes:
-		//  Measured:  `544`
-		//  Estimated: `83866`
-		// Minimum execution time: 63_403_000 picoseconds.
-		Weight::from_parts(64_420_000, 0)
-			.saturating_add(Weight::from_parts(0, 83866))
-			.saturating_add(T::DbWeight::get().reads(5))
-			.saturating_add(T::DbWeight::get().writes(4))
-	}
-	/// Storage: Referenda ReferendumInfoFor (r:1 w:1)
-	/// Proof: Referenda ReferendumInfoFor (max_values: None, max_size: Some(936), added: 3411, mode: MaxEncodedLen)
-	fn refund_decision_deposit() -> Weight {
-		// Proof Size summary in bytes:
-		//  Measured:  `384`
-		//  Estimated: `4401`
-		// Minimum execution time: 31_560_000 picoseconds.
-		Weight::from_parts(32_111_000, 0)
-			.saturating_add(Weight::from_parts(0, 4401))
-			.saturating_add(T::DbWeight::get().reads(1))
-			.saturating_add(T::DbWeight::get().writes(1))
-	}
-	/// Storage: Referenda ReferendumInfoFor (r:1 w:1)
-	/// Proof: Referenda ReferendumInfoFor (max_values: None, max_size: Some(936), added: 3411, mode: MaxEncodedLen)
-	fn refund_submission_deposit() -> Weight {
-		// Proof Size summary in bytes:
-		//  Measured:  `374`
-		//  Estimated: `4401`
-		// Minimum execution time: 31_536_000 picoseconds.
-		Weight::from_parts(32_118_000, 0)
-			.saturating_add(Weight::from_parts(0, 4401))
-			.saturating_add(T::DbWeight::get().reads(1))
-			.saturating_add(T::DbWeight::get().writes(1))
-	}
-	/// Storage: Referenda ReferendumInfoFor (r:1 w:1)
-	/// Proof: Referenda ReferendumInfoFor (max_values: None, max_size: Some(936), added: 3411, mode: MaxEncodedLen)
-	/// Storage: Scheduler Agenda (r:2 w:2)
-	/// Proof: Scheduler Agenda (max_values: None, max_size: Some(38963), added: 41438, mode: MaxEncodedLen)
-	fn cancel() -> Weight {
-		// Proof Size summary in bytes:
-		//  Measured:  `452`
-		//  Estimated: `83866`
-		// Minimum execution time: 39_132_000 picoseconds.
-		Weight::from_parts(39_878_000, 0)
-			.saturating_add(Weight::from_parts(0, 83866))
-			.saturating_add(T::DbWeight::get().reads(3))
-			.saturating_add(T::DbWeight::get().writes(3))
-	}
-	/// Storage: Referenda ReferendumInfoFor (r:1 w:1)
-	/// Proof: Referenda ReferendumInfoFor (max_values: None, max_size: Some(936), added: 3411, mode: MaxEncodedLen)
-	/// Storage: Scheduler Agenda (r:2 w:2)
-	/// Proof: Scheduler Agenda (max_values: None, max_size: Some(38963), added: 41438, mode: MaxEncodedLen)
-	/// Storage: Referenda MetadataOf (r:1 w:0)
-	/// Proof: Referenda MetadataOf (max_values: None, max_size: Some(52), added: 2527, mode: MaxEncodedLen)
-	fn kill() -> Weight {
-		// Proof Size summary in bytes:
-		//  Measured:  `693`
-		//  Estimated: `83866`
-		// Minimum execution time: 105_261_000 picoseconds.
-		Weight::from_parts(106_923_000, 0)
-			.saturating_add(Weight::from_parts(0, 83866))
-			.saturating_add(T::DbWeight::get().reads(4))
-			.saturating_add(T::DbWeight::get().writes(3))
-	}
-	/// Storage: Referenda TrackQueue (r:1 w:0)
-	/// Proof: Referenda TrackQueue (max_values: None, max_size: Some(2012), added: 4487, mode: MaxEncodedLen)
-	/// Storage: Referenda DecidingCount (r:1 w:1)
-	/// Proof: Referenda DecidingCount (max_values: None, max_size: Some(14), added: 2489, mode: MaxEncodedLen)
-	fn one_fewer_deciding_queue_empty() -> Weight {
-		// Proof Size summary in bytes:
-		//  Measured:  `207`
-		//  Estimated: `5477`
-		// Minimum execution time: 9_171_000 picoseconds.
-		Weight::from_parts(9_585_000, 0)
-			.saturating_add(Weight::from_parts(0, 5477))
-			.saturating_add(T::DbWeight::get().reads(2))
-			.saturating_add(T::DbWeight::get().writes(1))
-	}
-	/// Storage: Referenda TrackQueue (r:1 w:1)
-	/// Proof: Referenda TrackQueue (max_values: None, max_size: Some(2012), added: 4487, mode: MaxEncodedLen)
-	/// Storage: Referenda ReferendumInfoFor (r:1 w:1)
-	/// Proof: Referenda ReferendumInfoFor (max_values: None, max_size: Some(936), added: 3411, mode: MaxEncodedLen)
-	/// Storage: Balances InactiveIssuance (r:1 w:0)
-	/// Proof: Balances InactiveIssuance (max_values: Some(1), max_size: Some(16), added: 511, mode: MaxEncodedLen)
-	/// Storage: Scheduler Agenda (r:1 w:1)
-	/// Proof: Scheduler Agenda (max_values: None, max_size: Some(38963), added: 41438, mode: MaxEncodedLen)
-	fn one_fewer_deciding_failing() -> Weight {
-		// Proof Size summary in bytes:
-		//  Measured:  `3221`
-		//  Estimated: `42428`
-		// Minimum execution time: 49_135_000 picoseconds.
-		Weight::from_parts(50_860_000, 0)
-			.saturating_add(Weight::from_parts(0, 42428))
-			.saturating_add(T::DbWeight::get().reads(4))
-			.saturating_add(T::DbWeight::get().writes(3))
-	}
-	/// Storage: Referenda TrackQueue (r:1 w:1)
-	/// Proof: Referenda TrackQueue (max_values: None, max_size: Some(2012), added: 4487, mode: MaxEncodedLen)
-	/// Storage: Referenda ReferendumInfoFor (r:1 w:1)
-	/// Proof: Referenda ReferendumInfoFor (max_values: None, max_size: Some(936), added: 3411, mode: MaxEncodedLen)
-	/// Storage: Balances InactiveIssuance (r:1 w:0)
-	/// Proof: Balances InactiveIssuance (max_values: Some(1), max_size: Some(16), added: 511, mode: MaxEncodedLen)
-	/// Storage: Scheduler Agenda (r:1 w:1)
-	/// Proof: Scheduler Agenda (max_values: None, max_size: Some(38963), added: 41438, mode: MaxEncodedLen)
-	fn one_fewer_deciding_passing() -> Weight {
-		// Proof Size summary in bytes:
-		//  Measured:  `3221`
-		//  Estimated: `42428`
-		// Minimum execution time: 53_279_000 picoseconds.
-		Weight::from_parts(54_069_000, 0)
-			.saturating_add(Weight::from_parts(0, 42428))
-			.saturating_add(T::DbWeight::get().reads(4))
-			.saturating_add(T::DbWeight::get().writes(3))
-	}
-	/// Storage: Referenda ReferendumInfoFor (r:1 w:0)
-	/// Proof: Referenda ReferendumInfoFor (max_values: None, max_size: Some(936), added: 3411, mode: MaxEncodedLen)
-	/// Storage: Referenda TrackQueue (r:1 w:1)
-	/// Proof: Referenda TrackQueue (max_values: None, max_size: Some(2012), added: 4487, mode: MaxEncodedLen)
-	fn nudge_referendum_requeued_insertion() -> Weight {
-		// Proof Size summary in bytes:
-		//  Measured:  `3044`
-		//  Estimated: `5477`
-		// Minimum execution time: 22_537_000 picoseconds.
-		Weight::from_parts(23_853_000, 0)
-			.saturating_add(Weight::from_parts(0, 5477))
-			.saturating_add(T::DbWeight::get().reads(2))
-			.saturating_add(T::DbWeight::get().writes(1))
-	}
-	/// Storage: Referenda ReferendumInfoFor (r:1 w:0)
-	/// Proof: Referenda ReferendumInfoFor (max_values: None, max_size: Some(936), added: 3411, mode: MaxEncodedLen)
-	/// Storage: Referenda TrackQueue (r:1 w:1)
-	/// Proof: Referenda TrackQueue (max_values: None, max_size: Some(2012), added: 4487, mode: MaxEncodedLen)
-	fn nudge_referendum_requeued_slide() -> Weight {
-		// Proof Size summary in bytes:
-		//  Measured:  `3044`
-		//  Estimated: `5477`
-		// Minimum execution time: 22_686_000 picoseconds.
-		Weight::from_parts(23_947_000, 0)
-			.saturating_add(Weight::from_parts(0, 5477))
-			.saturating_add(T::DbWeight::get().reads(2))
-			.saturating_add(T::DbWeight::get().writes(1))
-	}
-	/// Storage: Referenda ReferendumInfoFor (r:1 w:1)
-	/// Proof: Referenda ReferendumInfoFor (max_values: None, max_size: Some(936), added: 3411, mode: MaxEncodedLen)
-	/// Storage: Referenda DecidingCount (r:1 w:0)
-	/// Proof: Referenda DecidingCount (max_values: None, max_size: Some(14), added: 2489, mode: MaxEncodedLen)
-	/// Storage: Referenda TrackQueue (r:1 w:1)
-	/// Proof: Referenda TrackQueue (max_values: None, max_size: Some(2012), added: 4487, mode: MaxEncodedLen)
-	fn nudge_referendum_queued() -> Weight {
-		// Proof Size summary in bytes:
-		//  Measured:  `3048`
-		//  Estimated: `5477`
-		// Minimum execution time: 28_373_000 picoseconds.
-		Weight::from_parts(29_033_000, 0)
-			.saturating_add(Weight::from_parts(0, 5477))
-			.saturating_add(T::DbWeight::get().reads(3))
-			.saturating_add(T::DbWeight::get().writes(2))
-	}
-	/// Storage: Referenda ReferendumInfoFor (r:1 w:1)
-	/// Proof: Referenda ReferendumInfoFor (max_values: None, max_size: Some(936), added: 3411, mode: MaxEncodedLen)
-	/// Storage: Referenda DecidingCount (r:1 w:0)
-	/// Proof: Referenda DecidingCount (max_values: None, max_size: Some(14), added: 2489, mode: MaxEncodedLen)
-	/// Storage: Referenda TrackQueue (r:1 w:1)
-	/// Proof: Referenda TrackQueue (max_values: None, max_size: Some(2012), added: 4487, mode: MaxEncodedLen)
-	fn nudge_referendum_not_queued() -> Weight {
-		// Proof Size summary in bytes:
-		//  Measured:  `3068`
-		//  Estimated: `5477`
-		// Minimum execution time: 28_137_000 picoseconds.
-		Weight::from_parts(28_716_000, 0)
-			.saturating_add(Weight::from_parts(0, 5477))
-			.saturating_add(T::DbWeight::get().reads(3))
-			.saturating_add(T::DbWeight::get().writes(2))
-	}
-	/// Storage: Referenda ReferendumInfoFor (r:1 w:1)
-	/// Proof: Referenda ReferendumInfoFor (max_values: None, max_size: Some(936), added: 3411, mode: MaxEncodedLen)
-	/// Storage: Scheduler Agenda (r:1 w:1)
-	/// Proof: Scheduler Agenda (max_values: None, max_size: Some(38963), added: 41438, mode: MaxEncodedLen)
-	fn nudge_referendum_no_deposit() -> Weight {
-		// Proof Size summary in bytes:
-		//  Measured:  `404`
-		//  Estimated: `42428`
-		// Minimum execution time: 25_880_000 picoseconds.
-		Weight::from_parts(26_405_000, 0)
-			.saturating_add(Weight::from_parts(0, 42428))
-			.saturating_add(T::DbWeight::get().reads(2))
-			.saturating_add(T::DbWeight::get().writes(2))
-	}
-	/// Storage: Referenda ReferendumInfoFor (r:1 w:1)
-	/// Proof: Referenda ReferendumInfoFor (max_values: None, max_size: Some(936), added: 3411, mode: MaxEncodedLen)
-	/// Storage: Scheduler Agenda (r:1 w:1)
-	/// Proof: Scheduler Agenda (max_values: None, max_size: Some(38963), added: 41438, mode: MaxEncodedLen)
-	fn nudge_referendum_preparing() -> Weight {
-		// Proof Size summary in bytes:
-		//  Measured:  `452`
-		//  Estimated: `42428`
-		// Minimum execution time: 26_349_000 picoseconds.
-		Weight::from_parts(27_181_000, 0)
-			.saturating_add(Weight::from_parts(0, 42428))
-			.saturating_add(T::DbWeight::get().reads(2))
-			.saturating_add(T::DbWeight::get().writes(2))
-	}
-	/// Storage: Referenda ReferendumInfoFor (r:1 w:1)
-	/// Proof: Referenda ReferendumInfoFor (max_values: None, max_size: Some(936), added: 3411, mode: MaxEncodedLen)
-	fn nudge_referendum_timed_out() -> Weight {
-		// Proof Size summary in bytes:
-		//  Measured:  `311`
-		//  Estimated: `4401`
-		// Minimum execution time: 17_735_000 picoseconds.
-		Weight::from_parts(18_130_000, 0)
-			.saturating_add(Weight::from_parts(0, 4401))
-			.saturating_add(T::DbWeight::get().reads(1))
-			.saturating_add(T::DbWeight::get().writes(1))
-	}
-	/// Storage: Referenda ReferendumInfoFor (r:1 w:1)
-	/// Proof: Referenda ReferendumInfoFor (max_values: None, max_size: Some(936), added: 3411, mode: MaxEncodedLen)
-	/// Storage: Referenda DecidingCount (r:1 w:1)
-	/// Proof: Referenda DecidingCount (max_values: None, max_size: Some(14), added: 2489, mode: MaxEncodedLen)
-	/// Storage: Balances InactiveIssuance (r:1 w:0)
-	/// Proof: Balances InactiveIssuance (max_values: Some(1), max_size: Some(16), added: 511, mode: MaxEncodedLen)
-	/// Storage: Scheduler Agenda (r:1 w:1)
-	/// Proof: Scheduler Agenda (max_values: None, max_size: Some(38963), added: 41438, mode: MaxEncodedLen)
-	fn nudge_referendum_begin_deciding_failing() -> Weight {
-		// Proof Size summary in bytes:
-		//  Measured:  `452`
-		//  Estimated: `42428`
-		// Minimum execution time: 36_244_000 picoseconds.
-		Weight::from_parts(37_174_000, 0)
-			.saturating_add(Weight::from_parts(0, 42428))
-			.saturating_add(T::DbWeight::get().reads(4))
-			.saturating_add(T::DbWeight::get().writes(3))
-	}
-	/// Storage: Referenda ReferendumInfoFor (r:1 w:1)
-	/// Proof: Referenda ReferendumInfoFor (max_values: None, max_size: Some(936), added: 3411, mode: MaxEncodedLen)
-	/// Storage: Referenda DecidingCount (r:1 w:1)
-	/// Proof: Referenda DecidingCount (max_values: None, max_size: Some(14), added: 2489, mode: MaxEncodedLen)
-	/// Storage: Balances InactiveIssuance (r:1 w:0)
-	/// Proof: Balances InactiveIssuance (max_values: Some(1), max_size: Some(16), added: 511, mode: MaxEncodedLen)
-	/// Storage: Scheduler Agenda (r:1 w:1)
-	/// Proof: Scheduler Agenda (max_values: None, max_size: Some(38963), added: 41438, mode: MaxEncodedLen)
-	fn nudge_referendum_begin_deciding_passing() -> Weight {
-		// Proof Size summary in bytes:
-		//  Measured:  `452`
-		//  Estimated: `42428`
-		// Minimum execution time: 38_250_000 picoseconds.
-		Weight::from_parts(38_771_000, 0)
-			.saturating_add(Weight::from_parts(0, 42428))
-			.saturating_add(T::DbWeight::get().reads(4))
-			.saturating_add(T::DbWeight::get().writes(3))
-	}
-	/// Storage: Referenda ReferendumInfoFor (r:1 w:1)
-	/// Proof: Referenda ReferendumInfoFor (max_values: None, max_size: Some(936), added: 3411, mode: MaxEncodedLen)
-	/// Storage: Balances InactiveIssuance (r:1 w:0)
-	/// Proof: Balances InactiveIssuance (max_values: Some(1), max_size: Some(16), added: 511, mode: MaxEncodedLen)
-	/// Storage: Scheduler Agenda (r:1 w:1)
-	/// Proof: Scheduler Agenda (max_values: None, max_size: Some(38963), added: 41438, mode: MaxEncodedLen)
-	fn nudge_referendum_begin_confirming() -> Weight {
-		// Proof Size summary in bytes:
-		//  Measured:  `505`
-		//  Estimated: `42428`
-		// Minimum execution time: 31_177_000 picoseconds.
-		Weight::from_parts(31_886_000, 0)
-			.saturating_add(Weight::from_parts(0, 42428))
-			.saturating_add(T::DbWeight::get().reads(3))
-			.saturating_add(T::DbWeight::get().writes(2))
-	}
-	/// Storage: Referenda ReferendumInfoFor (r:1 w:1)
-	/// Proof: Referenda ReferendumInfoFor (max_values: None, max_size: Some(936), added: 3411, mode: MaxEncodedLen)
-	/// Storage: Balances InactiveIssuance (r:1 w:0)
-	/// Proof: Balances InactiveIssuance (max_values: Some(1), max_size: Some(16), added: 511, mode: MaxEncodedLen)
-	/// Storage: Scheduler Agenda (r:1 w:1)
-	/// Proof: Scheduler Agenda (max_values: None, max_size: Some(38963), added: 41438, mode: MaxEncodedLen)
-	fn nudge_referendum_end_confirming() -> Weight {
-		// Proof Size summary in bytes:
-		//  Measured:  `488`
-		//  Estimated: `42428`
-		// Minimum execution time: 31_826_000 picoseconds.
-		Weight::from_parts(32_664_000, 0)
-			.saturating_add(Weight::from_parts(0, 42428))
-			.saturating_add(T::DbWeight::get().reads(3))
-			.saturating_add(T::DbWeight::get().writes(2))
-	}
-	/// Storage: Referenda ReferendumInfoFor (r:1 w:1)
-	/// Proof: Referenda ReferendumInfoFor (max_values: None, max_size: Some(936), added: 3411, mode: MaxEncodedLen)
-	/// Storage: Balances InactiveIssuance (r:1 w:0)
-	/// Proof: Balances InactiveIssuance (max_values: Some(1), max_size: Some(16), added: 511, mode: MaxEncodedLen)
-	/// Storage: Scheduler Agenda (r:1 w:1)
-	/// Proof: Scheduler Agenda (max_values: None, max_size: Some(38963), added: 41438, mode: MaxEncodedLen)
-	fn nudge_referendum_continue_not_confirming() -> Weight {
-		// Proof Size summary in bytes:
-		//  Measured:  `505`
-		//  Estimated: `42428`
-		// Minimum execution time: 28_957_000 picoseconds.
-		Weight::from_parts(29_810_000, 0)
-			.saturating_add(Weight::from_parts(0, 42428))
-			.saturating_add(T::DbWeight::get().reads(3))
-			.saturating_add(T::DbWeight::get().writes(2))
-	}
-	/// Storage: Referenda ReferendumInfoFor (r:1 w:1)
-	/// Proof: Referenda ReferendumInfoFor (max_values: None, max_size: Some(936), added: 3411, mode: MaxEncodedLen)
-	/// Storage: Balances InactiveIssuance (r:1 w:0)
-	/// Proof: Balances InactiveIssuance (max_values: Some(1), max_size: Some(16), added: 511, mode: MaxEncodedLen)
-	/// Storage: Scheduler Agenda (r:1 w:1)
-	/// Proof: Scheduler Agenda (max_values: None, max_size: Some(38963), added: 41438, mode: MaxEncodedLen)
-	fn nudge_referendum_continue_confirming() -> Weight {
-		// Proof Size summary in bytes:
-		//  Measured:  `509`
-		//  Estimated: `42428`
-		// Minimum execution time: 28_002_000 picoseconds.
-		Weight::from_parts(28_440_000, 0)
-			.saturating_add(Weight::from_parts(0, 42428))
-			.saturating_add(T::DbWeight::get().reads(3))
-			.saturating_add(T::DbWeight::get().writes(2))
-	}
-	/// Storage: Referenda ReferendumInfoFor (r:1 w:1)
-	/// Proof: Referenda ReferendumInfoFor (max_values: None, max_size: Some(936), added: 3411, mode: MaxEncodedLen)
-	/// Storage: Balances InactiveIssuance (r:1 w:0)
-	/// Proof: Balances InactiveIssuance (max_values: Some(1), max_size: Some(16), added: 511, mode: MaxEncodedLen)
-	/// Storage: Scheduler Agenda (r:2 w:2)
-	/// Proof: Scheduler Agenda (max_values: None, max_size: Some(38963), added: 41438, mode: MaxEncodedLen)
-	/// Storage: Scheduler Lookup (r:1 w:1)
-	/// Proof: Scheduler Lookup (max_values: None, max_size: Some(48), added: 2523, mode: MaxEncodedLen)
-	fn nudge_referendum_approved() -> Weight {
-		// Proof Size summary in bytes:
-		//  Measured:  `509`
-		//  Estimated: `83866`
-		// Minimum execution time: 43_527_000 picoseconds.
-		Weight::from_parts(44_536_000, 0)
-			.saturating_add(Weight::from_parts(0, 83866))
-			.saturating_add(T::DbWeight::get().reads(5))
-			.saturating_add(T::DbWeight::get().writes(4))
-	}
-	/// Storage: Referenda ReferendumInfoFor (r:1 w:1)
-	/// Proof: Referenda ReferendumInfoFor (max_values: None, max_size: Some(936), added: 3411, mode: MaxEncodedLen)
-	/// Storage: Balances InactiveIssuance (r:1 w:0)
-	/// Proof: Balances InactiveIssuance (max_values: Some(1), max_size: Some(16), added: 511, mode: MaxEncodedLen)
-	/// Storage: Scheduler Agenda (r:1 w:1)
-	/// Proof: Scheduler Agenda (max_values: None, max_size: Some(38963), added: 41438, mode: MaxEncodedLen)
-	fn nudge_referendum_rejected() -> Weight {
-		// Proof Size summary in bytes:
-		//  Measured:  `505`
-		//  Estimated: `42428`
-		// Minimum execution time: 31_767_000 picoseconds.
-		Weight::from_parts(32_407_000, 0)
-			.saturating_add(Weight::from_parts(0, 42428))
-			.saturating_add(T::DbWeight::get().reads(3))
-			.saturating_add(T::DbWeight::get().writes(2))
-	}
-	/// Storage: Referenda ReferendumInfoFor (r:1 w:0)
-	/// Proof: Referenda ReferendumInfoFor (max_values: None, max_size: Some(936), added: 3411, mode: MaxEncodedLen)
-	/// Storage: Preimage StatusFor (r:1 w:0)
-	/// Proof: Preimage StatusFor (max_values: None, max_size: Some(91), added: 2566, mode: MaxEncodedLen)
-	/// Storage: Referenda MetadataOf (r:0 w:1)
-	/// Proof: Referenda MetadataOf (max_values: None, max_size: Some(52), added: 2527, mode: MaxEncodedLen)
-	fn set_some_metadata() -> Weight {
-		// Proof Size summary in bytes:
-		//  Measured:  `455`
-		//  Estimated: `4401`
-		// Minimum execution time: 21_013_000 picoseconds.
-		Weight::from_parts(21_503_000, 0)
-			.saturating_add(Weight::from_parts(0, 4401))
-			.saturating_add(T::DbWeight::get().reads(2))
-			.saturating_add(T::DbWeight::get().writes(1))
-	}
-	/// Storage: Referenda ReferendumInfoFor (r:1 w:0)
-	/// Proof: Referenda ReferendumInfoFor (max_values: None, max_size: Some(936), added: 3411, mode: MaxEncodedLen)
-	/// Storage: Referenda MetadataOf (r:1 w:1)
-	/// Proof: Referenda MetadataOf (max_values: None, max_size: Some(52), added: 2527, mode: MaxEncodedLen)
-	fn clear_metadata() -> Weight {
-		// Proof Size summary in bytes:
-		//  Measured:  `388`
-		//  Estimated: `4401`
-		// Minimum execution time: 18_535_000 picoseconds.
-		Weight::from_parts(19_056_000, 0)
-			.saturating_add(Weight::from_parts(0, 4401))
-			.saturating_add(T::DbWeight::get().reads(2))
-			.saturating_add(T::DbWeight::get().writes(1))
-	}
-}
